requirements.txt
netmiko>=4.4.0
rich==13.7.0
click==8.1.7
PyYAML==6.0.1
